Error messages and override conditions
Created the section which deals with error messages and override conditions.

This does a bunch of IF statements to change the vb (verb number) if neccesary, and change the message waiting to be output to the user, preparing for branching to a bunch of subroutines (which we will need to do as ELIF statements)
Had to put a loop in around the main calls which condition is returned from the error/override routine (and in turn from branching routine) to get around the lack of GOTOs.