Mainly worked on Player() - now responds to verbs - and changed direction system.
Added look, showMessages, move, getInstruction, inputSplit functions to Player. Player can now check for  and respond to verbs. Added dummy verb LLAMA to test, and verb E to allow movement East. Which works!
Changed directions in Room to be a dictionary so as to be able to move to arbitrary rooms rather than performing calculations as in the original. A little more work to set up (and I've commented a load out for now!) but far easier to actually use. A few tweaks for the main loop to show messages and 'look' at each round.